refactor(broker): move query api into separate package
